% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_funs.R
\name{applyIQRrule}
\alias{applyIQRrule}
\title{Apply the 1.5 x IQR Rule to a vector}
\usage{
applyIQRrule(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a numeric vector with outliers removed
}
\description{
Apply the 1.5 x IQR Rule to a vector
}
