% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{are_unique}
\alias{are_unique}
\title{Are the elements of the vector unique?}
\usage{
are_unique(vec)
}
\arguments{
\item{vec}{a vector}
}
\value{
binary
}
\description{
Are the elements of the vector unique?
}
\examples{
are_unique(c(1,1,2,3))
}
