% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{beforeParenthesis}
\alias{beforeParenthesis}
\title{Pulls out disease names and formats to column names}
\usage{
beforeParenthesis(vec)
}
\arguments{
\item{vec}{}
}
\value{
vec
}
\description{
Pulls out disease names and formats to column names
}
