% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_funs.R
\name{bio_applyInstanceNames}
\alias{bio_applyInstanceNames}
\title{Takes a raw UKB data table and applies instance names to columns
Only for use where one column per instance, can supply a textTag}
\usage{
bio_applyInstanceNames(df, textTag = "Inst")
}
\arguments{
\item{df}{raw UKB data.frame}

\item{textTag}{optional tag to show the content of the field}
}
\value{
a data.frame with new names
}
\description{
Takes a raw UKB data table and applies instance names to columns
Only for use where one column per instance, can supply a textTag
}
