% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_funs.R
\name{bio_numeric}
\alias{bio_numeric}
\title{Converts all fields except the first into numbers}
\usage{
bio_numeric(df, colIndex = 2:ncol(df))
}
\arguments{
\item{df}{a data.frame}

\item{colIndex}{columns to convert, defaults to all but first}
}
\value{
a data.frame
}
\description{
Converts all fields except the first into numbers
}
