% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_funs.R
\name{bio_numFun}
\alias{bio_numFun}
\title{applies numFun to all but first col in dataframe
removes negative special codes, and makes the rest numeric}
\usage{
bio_numFun(df, colIndex = 2:ncol(df))
}
\arguments{
\item{df}{a data.frame}

\item{colIndex}{columns to convert, defaults to all but first}
}
\value{
a data.frame
}
\description{
applies numFun to all but first col in dataframe
removes negative special codes, and makes the rest numeric
}
