% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_funs.R
\name{bio_read}
\alias{bio_read}
\title{Main Current UKB reading function}
\usage{
bio_read(udi, baseOnly = FALSE, printHead = FALSE)
}
\arguments{
\item{udi}{UKB field number}

\item{baseOnly}{binary, restrict result to baseline only?}

\item{printHead}{binary, print the first part of input data?}
}
\value{
a data.frame with text contents
}
\description{
Main Current UKB reading function
}
