% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_funs.R
\name{bio_rename}
\alias{bio_rename}
\title{Rename all but first column
Useful for changing a single name in a tidyverse sentence}
\usage{
bio_rename(sett, namevec)
}
\arguments{
\item{sett}{data.frame}

\item{namevec}{new names, must be length= ncol(sett)-1}
}
\value{
renamed data.frame
}
\description{
Rename all but first column
Useful for changing a single name in a tidyverse sentence
}
