% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_funs.R
\name{biochem_read}
\alias{biochem_read}
\title{Convenience function for aggregating the mean for biochemistry records}
\usage{
biochem_read(udi, thename)
}
\arguments{
\item{udi}{UKB field number}

\item{thename}{character string, name for the new aggregated field}
}
\value{
a numeric data.frame
}
\description{
Convenience function for aggregating the mean for biochemistry records
}
