% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_age.R
\name{calc_time_since}
\alias{calc_time_since}
\title{Calculate age}
\usage{
calc_time_since(
  time,
  age_day = lubridate::today(),
  units = "years",
  floor = TRUE
)
}
\arguments{
\item{time}{date-of-birth, the day to start calculating age.}

\item{age_day}{the date on which age is to be calculated.}

\item{units}{unit to measure age in. Defaults to \code{"years"}. Passed to \link{\code{duration}}.}

\item{floor}{boolean for whether or not to floor the result. Defaults to \code{TRUE}.}
}
\value{
Age in \code{units}. Will be an integer if \code{floor = TRUE}.
}
\description{
By default, calculates the typical "age in years", with a
\code{floor} applied so that you are, e.g., 5 years old from
5th birthday through the day before your 6th birthday. Set
\code{floor = FALSE} to return decimal ages, and change \code{units}
for units other than years.
}
\note{
taken from here \url{https://stackoverflow.com/questions/27096485/change-a-column-from-birth-date-to-age-in-r}
}
