% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{checkkey}
\alias{checkkey}
\title{Check if the keys in a data frame are unique / good to join
Specificallly for UKB, the key field is assumed to be f.eid}
\usage{
checkkey(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
binary
}
\description{
Check if the keys in a data frame are unique / good to join
Specificallly for UKB, the key field is assumed to be f.eid
}
\examples{
checkkey(df)
}
