% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{collapseString}
\alias{collapseString}
\title{Collapse a vector of values into a single string
separated by commas}
\usage{
collapseString(x, distinctVals = FALSE)
}
\arguments{
\item{x}{a vector}

\item{distinctVals}{binary, only retain unique values?}
}
\value{
a string
}
\description{
Collapse a vector of values into a single string
separated by commas
}
