% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_funs.R
\name{do_LOG}
\alias{do_LOG}
\title{Create a log field with sources as input}
\usage{
do_LOG(temp, sources)
}
\arguments{
\item{temp}{data.frame with binary indicators}

\item{sources}{character vector showing source of indicator variables}
}
\value{
a string vector, string result for each row
}
\description{
Create a log field with sources as input
}
