% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_covars.R
\name{extract_covars}
\alias{extract_covars}
\title{extract UKBB covaraties}
\usage{
extract_covars(dt_pheno, dt_sample_qc, keep = "eid")
}
\arguments{
\item{dt_pheno}{data.table for phenotypes}

\item{dt_sample_qc}{data.table for sample QC}

\item{keep}{what columns should be kept from dt_pheno file?}
}
\description{
a function that extracts basisc UKBB parameters
}
