% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{fixNAs}
\alias{fixNAs}
\title{Convert missing values in a vector into zeros}
\usage{
fixNAs(thevec)
}
\arguments{
\item{thevec}{}
}
\value{
numeric vector
}
\description{
Convert missing values in a vector into zeros
}
\examples{
IP$diabetes<- fixNAs(IP$diabetes)
}
