% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{fromSrc}
\alias{fromSrc}
\title{Points a read function towards the nominated data source
Relies on prior specification of the dataSourceFolder}
\usage{
fromSrc(filename)
}
\arguments{
\item{filename}{character string, the name of the file}
}
\value{
extended file path string
}
\description{
Points a read function towards the nominated data source
Relies on prior specification of the dataSourceFolder
}
