% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{howmany}
\alias{howmany}
\title{Returns the number of unique values in the vector}
\usage{
howmany(thevar)
}
\arguments{
\item{thevar}{a vector}
}
\value{
an integer
}
\description{
Returns the number of unique values in the vector
}
