% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd10chapters}
\alias{icd10chapters}
\title{International Classification of Diseases Revision 10 (ICD-10) chapters}
\format{An object of class \code{data.frame} with 21 rows and 3 columns.}
\usage{
icd10chapters
}
\description{
A dataset containing the ICD-10 chapter titles - a top level description of diagnoses classes (or blocks)
}
\keyword{datasets}
