% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd10codes}
\alias{icd10codes}
\title{International Classification of Diseases Revision 10 (ICD-10) codes}
\format{
An object of class \code{data.frame} with 18761 rows and 2 columns.
}
\usage{
icd10codes
}
\description{
A dataset containing the full set ICD-10 diagnoses
}
\keyword{datasets}
