% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd9chapters}
\alias{icd9chapters}
\title{International Classification of Diseases Revision 9 (ICD-9) chapters}
\format{An object of class \code{data.frame} with 19 rows and 3 columns.}
\usage{
icd9chapters
}
\description{
A dataset containing the ICD-9 chapter titles - a top level description of diagnoses classes (or blocks)
}
\keyword{datasets}
