% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd9codes}
\alias{icd9codes}
\title{International Classification of Diseases Revision 9 (ICD-9) codes}
\format{An object of class \code{data.frame} with 13679 rows and 2 columns.}
\usage{
icd9codes
}
\description{
A dataset containing the full set ICD-9 diagnoses
}
\keyword{datasets}
