% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis.R
\name{icd_code}
\alias{icd_code}
\title{Retrieves description for a ICD code.}
\usage{
icd_code(icd.code, icd.version)
}
\arguments{
\item{icd.code}{The ICD diagnosis code to be looked up.}

\item{icd.version}{The ICD version (or revision) number, 9 or 10.}
}
\description{
Retrieves description for a ICD code.
}
