% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{label_cols}
\alias{label_cols}
\title{Add labels to column names (all but first)}
\usage{
label_cols(sett, label, before = "before")
}
\arguments{
\item{sett}{data.frame}

\item{label}{character string}

\item{before}{placement of label c("before","after")}
}
\value{
data.frame with new labels
}
\description{
Add labels to column names (all but first)
}
