% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lookup}
\alias{lookup}
\title{Create a lookup table}
\usage{
lookup(data, key, value)
}
\arguments{
\item{data}{A dataframe.}

\item{key}{Name of column to use as key.}

\item{value}{Name of column to use for value.}
}
\description{
The lookup table is a named vector
}
