% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_icd_mapping.R
\name{make_icd_mapping}
\alias{make_icd_mapping}
\title{make icd mapping funcs}
\usage{
make_icd_mapping(
  from,
  to,
 
    icd_path = "/well/lindgren/UKBIOBANK/ferreira/phenotype_extraction/phenotype_lists/ICD10_codes_converted_ICD9_v2v3_codes.txt"
)
}
\arguments{
\item{from}{a string}

\item{to}{a string}

\item{icd_path}{path to icd bridge file}
}
\description{
Create a function that allows quick mapping between ICD and disease cateogories
}
\examples{
\dontrun{
f <- create_icd_mapping('ICD10','disease')
f(c('124','Z851'))
}
}
