% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multiplot}
\alias{multiplot}
\title{Draws a grid of ggplot figures}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\description{
For arranging multiple ggplot2 plots in a grid. Found in the \href{http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/}{Cookbook for R}.
}
