% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{newmatrix}
\alias{newmatrix}
\title{Create an empty data frame}
\usage{
newmatrix(nr, nc, thing = NA)
}
\arguments{
\item{nr}{integer, number of rows desired}

\item{nc}{integer, number of columns desired}

\item{thing}{value to populate the new df, defaults to NA}
}
\value{
a data frame
}
\description{
Create an empty data frame
}
