% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{niceNames}
\alias{niceNames}
\title{Converts a bunch of strings to sensible column names}
\usage{
niceNames(vec)
}
\arguments{
\item{vec}{}
}
\value{
vec
}
\description{
Converts a bunch of strings to sensible column names
}
