% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{not_including}
\alias{not_including}
\title{Exclude elements in a vector}
\usage{
not_including(x, elements)
}
\arguments{
\item{x}{a vector}

\item{elements}{a single element or several elements}
}
\value{
x vector not including elements
}
\description{
Exclude elements in a vector
}
\examples{
c("a", "b", "c", "d") \%>\% not_including(c("a","b"))
}
