% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_funs.R
\name{numFun}
\alias{numFun}
\title{Convert to numeric and change negative numbers to NA}
\usage{
numFun(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a numeric vector
}
\description{
Convert to numeric and change negative numbers to NA
}
