% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{padded}
\alias{padded}
\title{Adds zero padding to the left}
\usage{
padded(thingToPad, width)
}
\arguments{
\item{thingToPad}{vector}

\item{width}{integer - intended final width of elements}
}
\value{
vector
}
\description{
Adds zero padding to the left
}
\examples{
nice_numbers<- padded(c(1,2,3,20), 2)
}
