% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{percentable}
\alias{percentable}
\title{Simple percentage table of a vector}
\usage{
percentable(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a percentage table
}
\description{
Simple percentage table of a vector
}
