% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pull}
\alias{pull}
\title{Collects a column from a tidyverse tibble as an atomic vector.}
\usage{
pull(data, y)
}
\arguments{
\item{data}{A tidyverse tibble, tbl_df}

\item{y}{The column to be collected. Can be a numeric index or the column
name (quoted or not)}
}
\description{
A function found in a
\href{http://stackoverflow.com/a/24730843/3819656}{Stack Overflow} discussion
to retrieve a  tibble column as a vector
}
