% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{putFolder}
\alias{putFolder}
\title{Inserts a new folder into the current directory
If the folder already exists it does nothing}
\usage{
putFolder(folderName)
}
\arguments{
\item{folderName}{character string}
}
\value{
none
}
\description{
Inserts a new folder into the current directory
If the folder already exists it does nothing
}
\examples{
putFolder("Graphics")
}
