% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{putZeros}
\alias{putZeros}
\title{Replace NAs in an entire data frame with zeros}
\usage{
putZeros(OB)
}
\arguments{
\item{OB}{data.frame}
}
\value{
data.frame
}
\description{
Replace NAs in an entire data frame with zeros
}
\examples{
OB<- OB \%>\% putZeros()
}
