% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{shownames}
\alias{shownames}
\title{Show all names in a data frame with partial match
to a TEXT tag}
\usage{
shownames(SETT, TEXT)
}
\arguments{
\item{SETT}{data.frame}

\item{TEXT}{string}
}
\value{
character vector of column name matches
}
\description{
Show all names in a data frame with partial match
to a TEXT tag
}
