% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{to_clipboard}
\alias{to_clipboard}
\title{Copies object x to the clipboard}
\usage{
to_clipboard(x = .Last.value)
}
\arguments{
\item{x}{a vector or data.frame}
}
\value{
none
}
\description{
Copies object x to the clipboard
}
\examples{
to_clipboard(t0)
}
