% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukb_dataset.R
\name{ukb_df_join_all}
\alias{ukb_df_join_all}
\title{Recursively join a list of UKB datasets}
\usage{
ukb_df_join_all(..., by = "eid")
}
\arguments{
\item{...}{Supply unquoted names of to-be-merged UKB datasets created with \code{\link{ukb_df}}. Arguments are passed to \code{list}, and then to the \code{plyr::join_all} argument \code{dfs}, which takes a list of dataframes.}

\item{by}{Variable used to merge multiple dataframes (default = "eid").}
}
\description{
A thin wrapper around \code{plyr::join_all} to merge multiple UKB datasets.
}
\examples{
\dontrun{
# If you have multiple UKB filesets, tidy then merge them.

ukb1234_data <- ukb_df("ukb1234")
ukb2345_data <- ukb_df("ukb2345")
ukb3456_data <- ukb_df("ukb3456")

my_ukb_data <- ukb_df_join_all(ukb1234_data, ukb2345_data, ukb3456_data)
}

}
