% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{ukb_gen_pcs}
\alias{ukb_gen_pcs}
\title{Genetic principal components}
\usage{
ukb_gen_pcs(data)
}
\arguments{
\item{data}{A UKB dataset created with \code{\link{ukb_df}}.}
}
\description{
\lifecycle{defunct}

 These are the principal components derived on the UK Biobank subsample with interim genotype data. UKB have published \href{http://www.ukbiobank.ac.uk/wp-content/uploads/2014/04/UKBiobank_genotyping_QC_documentation-web.pdf}{full details of genotyping and quality control} for the interim genotype data.
}
