% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{ukb_gen_plinkid}
\alias{ukb_gen_plinkid}
\title{Two-column ID for plink individual subset selection}
\usage{
ukb_gen_plinkid(file, id)
}
\arguments{
\item{file}{Filename to create disk (including path if not current directory).}

\item{id}{A character vector of IDs to save to file for plink keep/remove subsetting.}
}
\description{
Writes a two-column space-delimited text file of IDs (without header) for sample subset selection in \code{\href{https://www.cog-genomics.org/plink2/filter#snp}{plink}}, using \code{--keep} or \code{--remove}.
}
