% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwrite.R
\name{ukb_gen_read_sample}
\alias{ukb_gen_read_sample}
\title{Reads an Oxford format sample file}
\usage{
ukb_gen_read_sample(file, col.names = c("id_1", "id_2", "missing"),
  row.skip = 2)
}
\arguments{
\item{file}{A path to a sample file.}

\item{col.names}{A character vector of column names. Default: c("id_1", "id_2", "missing")}

\item{row.skip}{Number of lines to skip before reading data.}
}
\description{
This is a wrapper for \code{read_table} that reads an Oxford format .sample file. If you use the unedited sample file as supplied with your genetic data, you should aonly need to specifiy the first argument, file.
}
\seealso{
\code{\link{ukb_gen_read_fam}} to read a fam file
}
