% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwrite.R
\name{ukb_gen_write_bgenie}
\alias{ukb_gen_write_bgenie}
\title{Writes a BGENIE format phenotype or covariate file.}
\usage{
ukb_gen_write_bgenie(x, ukb.sample, path, ukb.variables, ukb.id = "eid",
  na.strings = "-999")
}
\arguments{
\item{x}{A data frame to write to disk.}

\item{ukb.sample}{Path to the UKB sample file.}

\item{path}{A path to a file.}

\item{ukb.variables}{A character vector of either the phenotypes for a BGENIE phenotype file, or covariates for a BGENIE covariate file.}

\item{ukb.id}{The eid variable name (default = "eid").}
}
\description{
Writes a space-delimited file with a header, with missing character set to "-999", and observations (i.e. UKB subject ids) in sample file order.
}
\details{
See [BGENIE usage](https://jmarchini.org/bgenie-usage/) for descriptions of the \code{--pheno} and \code{--covar} flags to read phenotype and covariate data into BGENIE.
}
\seealso{
\code{\link{ukb_gen_read_sample}} to read a sample file, and \code{\link{ukb_gen_write_plink}} to write phenotype and covariate files to plink format.
}
