% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwrite.R
\name{ukb_gen_write_plink}
\alias{ukb_gen_write_plink}
\title{Writes a plink format phenotype or covariate file}
\usage{
ukb_gen_write_plink(x, path, ukb.variables, ukb.id = "eid")
}
\arguments{
\item{x}{A data frame to write to disk.}

\item{path}{A path to a file.}

\item{ukb.variables}{A character vector of either the phenotypes for a plink phenotype file, or covariates for a plink covariate file.}

\item{ukb.id}{The eid variable name (default = "eid").}
}
\description{
This function writes a tab-delimited file with header, with the obligatory first two columns FID and IID.
}
\details{
Use the \code{--pheno-name} and \code{--covar-name} plink flags to select columns by name. See the plink documentation for \code{--pheno}, \code{--mpheno}, \code{--pheno-name}, and \code{--covar}, \code{--covar-name}, \code{--covar-number} flags.
}
\seealso{
\code{\link{ukb_gen_read_sample}} to read a sample file, and \code{\link{ukb_gen_write_bgenie}} to write phenotype and covariate files to BGENIE format.
}
