% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis.R
\name{ukb_icd_chapter}
\alias{ukb_icd_chapter}
\title{Displays a table for the International Classification of Diseases (ICD)}
\usage{
ukb_icd_chapter(icd.version)
}
\arguments{
\item{icd.version}{The ICD icd.version (or revision) number, 9 or 10.}
}
\description{
Displays a table for the International Classification of Diseases (ICD)
}
