% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis.R
\name{ukb_icd_code_meaning}
\alias{ukb_icd_code_meaning}
\title{Retrieves description for a ICD code.}
\usage{
ukb_icd_code_meaning(icd.code, icd.version)
}
\arguments{
\item{icd.code}{The ICD diagnosis code to be looked up.}

\item{icd.version}{The ICD version (or revision) number, 9 or 10.}
}
\description{
Retrieves description for a ICD code.
}
\seealso{
\code{\link{ukb_icd_diagnosis}}, \code{\link{ukb_icd_keyword}}, \code{\link{ukb_icd_prevalence}}
}
