% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis.R
\name{ukb_icd_description}
\alias{ukb_icd_description}
\title{Retrieves diagnoses containing a description.}
\usage{
ukb_icd_description(icd.version, description)
}
\arguments{
\item{icd.version}{The ICD version (or revision) number, 9 or 10.}

\item{description}{A regular expression to be looked up in the ICD descriptions, e.g., "cardiovascular"}
}
\description{
Retrieves diagnoses containing a description.
}
