% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_util.R
\name{ukb_md5}
\alias{ukb_md5}
\alias{ukb_unpack}
\alias{ukb_conv}
\alias{ukb_fetch_bulk}
\title{UKB MD5 Checksum}
\usage{
ukb_md5(file, ...)

ukb_unpack(file, key, ...)

ukb_conv(
  file,
  encoding_file = "encoding.ukb",
  type = c("r", "docs", "csv", "sas", "stata", "lims", "bulk", "txt"),
  ...
)

ukb_fetch_bulk(file, key, start = NULL, outdir = NULL, ...)
}
\arguments{
\item{file}{name of file to run utility on}

\item{...}{additional arguments to pass to
\code{\link{ukb_util_path}}}

\item{key}{file to key to unpack/decrypt file}

\item{encoding_file}{encoding file to map for `ukbconv`.  If want no
encoding, set to \code{NULL}}

\item{type}{type of conversion to do}

\item{start}{start of the fetching, 1-indexed}

\item{outdir}{output directory of download}
}
\value{
A character string
}
\description{
UKB MD5 Checksum
}
