% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filehandlers.R
\name{ukb_util_conv}
\alias{ukb_util_conv}
\title{Converts unpacked UKB data to other formats.}
\usage{
ukb_util_conv(
  file,
  type = c("r", "docs", "csv", "sas", "stata", "lims", "bulk", "txt"),
  encoding_file = "encoding.ukb",
  ...
)
}
\arguments{
\item{file}{Path to decrypted file to convert.}

\item{type}{Type of conversion to do.}

\item{encoding_file}{encoding file to map for `ukbconv`.  If want no
encoding, set to \code{NULL}}

\item{...}{Additional arguments to pass to
\code{\link{ukb_util_get}}.}
}
\description{
Converts unpacked UKB data to other formats.
}
