% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filehandlers.R
\name{ukb_util_encoding}
\alias{ukb_util_encoding}
\title{Downloads encoding dictionaries for use with ukb_util_conv.}
\usage{
ukb_util_encoding(out_dir = tempdir())
}
\arguments{
\item{out_dir}{The output directory to download the UKB utility to. Default \code{tempdir()}.}
}
\description{
Downloads encoding dictionaries for use with ukb_util_conv.
}
