% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filehandlers.R
\name{ukb_util_fetch}
\alias{ukb_util_fetch}
\title{Downloads approved bulk data files.}
\usage{
ukb_util_fetch(file, key, start = NULL, out_dir = NULL, ...)
}
\arguments{
\item{file}{Path to bulk file}

\item{key}{Path to key file.}

\item{start}{start of the fetching, 1-indexed}

\item{out_dir}{output directory of download}

\item{...}{Additional arguments to pass to
\code{\link{ukb_util_get}}.}
}
\description{
Downloads approved bulk data files.
}
