% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filehandlers.R
\name{ukb_util_get}
\alias{ukb_util_get}
\title{Downloads individual UKB utlities and file handlers.}
\usage{
ukb_util_get(
  util = c("ukbmd5", "ukbconv", "ukbunpack", "ukbfetch", "ukblink", "ukbgene",
    "encoding.ukb"),
  download = TRUE,
  out_dir = tempdir()
)
}
\arguments{
\item{util}{Name of the utility. Must be one of \code{"ukbmd5"}, \code{"ukbconv"}, \code{"ukbunpack"}, \code{"ukbfetch"}, \code{"ukblink"}, \code{"ukbgene"}, \code{"encoding.ukb"}. For a description of the UKB utilies and file handlers, see \href{http://biobank.ndph.ox.ac.uk/showcase/download.cgi}{UKB Downloads}.}

\item{download}{Should the utility be downloaded if not found? Default is \code{TRUE}.}

\item{out_dir}{The output directory to download the UKB utility to. Default \code{tempdir()}.}
}
\value{
Path to the downloaded utility.
}
\description{
Downloads individual UKB utlities and file handlers.
}
\examples{
md5 = ukb_util_get("ukbmd5")
file.remove(md5)
}
