% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filehandlers.R
\name{ukb_util_md5}
\alias{ukb_util_md5}
\title{Calculates size and MD5 of a UKB utlity file.}
\usage{
ukb_util_md5(file, ...)
}
\arguments{
\item{file}{Path to file to run MD5 utility on.}

\item{...}{Additional arguments to pass to
\code{\link{ukb_util_get}}.}
}
\value{
A character string
}
\description{
Calculates size and MD5 of a UKB utlity file.
}
