% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_util.R
\name{ukb_util_path}
\alias{ukb_util_path}
\alias{ukb_encoding}
\title{Get Path to UKB Utilitiy}
\usage{
ukb_util_path(
  util = c("ukbmd5", "ukbconv", "ukbunpack", "ukbfetch", "ukblink", "ukbgene",
    "encoding.ukb"),
  download = TRUE,
  outdir = tempdir()
)

ukb_encoding(outdir = tempdir())
}
\arguments{
\item{util}{Name of the utility}

\item{download}{Should the utility be downloaded if not found?}

\item{outdir}{The output directory to download the utility}
}
\value{
A path to the utility
}
\description{
Get Path to UKB Utilitiy
}
\examples{
ukb_util_path("ukbmd5")
}
