% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filehandlers.R
\name{ukb_util_unpack}
\alias{ukb_util_unpack}
\title{Unpacks (decrypts and decompresses) UKB data.}
\usage{
ukb_util_unpack(file, key, ...)
}
\arguments{
\item{file}{Path to file to unpack/decrypt.}

\item{key}{file to key to unpack/decrypt file}

\item{...}{Additional arguments to pass to
\code{\link{ukb_util_get}}.}
}
\description{
Unpacks (decrypts and decompresses) UKB data.
}
