% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ukbcentre}
\alias{ukbcentre}
\title{UKB assessment centre}
\format{An object of class \code{data.frame} with 24 rows and 2 columns.}
\usage{
ukbcentre
}
\description{
A dataset containing the 22 assessment centres (as well as pilot test centre and a revisit centre)
}
\keyword{datasets}
