% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{ukb_context}
\alias{ukb_context}
\title{Demographics of a sample subset, or an individual}
\usage{
ukb_context(data, comparison.var, sample.sub, sample.ref)
}
\arguments{
\item{data}{A UKB dataset.}

\item{sample.sub}{A vector of UKB ids, or logical or row index, that defines the subset of interest.}

\item{sample.ref}{A vector of UKB ids (or a logical or row index), that defines an alternative reference population. By default the reference is the full sample.}
}
\description{
Describes a subset of UKB sample, or an individual, relative to the full UK Biobank sample, or a specified UKB subset. This function is intended as an exploratory data analysis and quality control tool, and as such only provides summary statistics and graphical context for individual's data.
}
