% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetics_qc.R
\name{ukb_gen_samples_to_remove}
\alias{ukb_gen_samples_to_remove}
\title{Related samples (with data on the variable of interest) to remove}
\usage{
ukb_gen_samples_to_remove(data, ukb_with_data, cutoff = 0.0884)
}
\arguments{
\item{data}{The UKB relatedness data as a dataframe (header: ID1, ID2, HetHet, IBS0, Kinship)}

\item{ukb_with_data}{A character vector of ukb eids with data on the phenotype of interest}

\item{cutoff}{KING kingship coefficient cutoff (default 0.0884 includes pairs with greater than 3rd-degree relatedness)}
}
\value{
A dataframe (header: ID1, ID2, HetHet, IBS0, Kinship) for the subset of individuals with data.
}
\description{
Related samples (with data on the variable of interest) to remove
}
