% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwrite.R
\name{ukb_gen_write_bgenie}
\alias{ukb_gen_write_bgenie}
\title{Writes a BGENIE format phenotype or covariate file.}
\usage{
ukb_gen_write_bgenie(x, ukb.sample, path, ukb.variables, ukb.id = "eid",
  na.strings = "-999")
}
\arguments{
\item{x}{A UKB dataset.}

\item{ukb.sample}{Path to the UKB sample file.}

\item{path}{A path to a file.}

\item{ukb.variables}{A character vector of either the phenotypes for a BGENIE phenotype file, or covariates for a BGENIE covariate file.}

\item{ukb.id}{The eid variable name (default = "eid").}

\item{na.strings}{Characer string to be used for missing value in output file. Default = "-999"}
}
\description{
Writes a space-delimited file with a header, missing character set to "-999", and observations (i.e. UKB subject ids) in sample file order. Use this function to write phenotype and covariate files for downstream genetic analysis in \href{https://jmarchini.org/bgenie/}{BGENIE} - the format is the same.
}
\details{
See href{https://jmarchini.org/bgenie-usage/}{BGENIE usage} for descriptions of the \code{--pheno} and \code{--covar} flags to read phenotype and covariate data into BGENIE.
}
\seealso{
\code{\link{ukb_gen_read_sample}} to read a sample file, \code{\link{ukb_gen_excl_to_na}} to update a phenotype with NAs for samples to-be-excluded based on genetic metadata, and \code{\link{ukb_gen_write_plink}} to write phenotype and covariate files to PLINK format.
}
