% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwrite.R
\name{ukb_gen_write_plink_excl}
\alias{ukb_gen_write_plink_excl}
\title{Writes a plink format file for combined exclusions}
\usage{
ukb_gen_write_plink_excl(ukb.path)
}
\arguments{
\item{ukb.path}{}
}
\description{
Writes a combined exclusions file including UKB recommended exclusions, heterozygosity exclusions (+/- 3*sd from mean), genetic ethnicity exclusions (based on the UKB genetic ethnic grouping variable, field 1002), and relatedness exclusions (a randomly-selected member of each related pair). For exclusion of individuals from a genetic analysis, the plink flag \code{--remove} accepts a space/tab-delimited text file with family IDs in the first column and within-family IDs in the second column (i.e., FID IID), without a header.
}
\seealso{
\link{\code{ukb_gen_meta}}, \link{\code{ukb_gen_pcs}} which retrieve variables to be included in a covariate file. \link{\code{ukb_gen_excl_to_na}} to update a phenotype with NAs for samples to-be-excluded based on genetic metadata, and \link{\code{ukb_gen_write_plink}} and \link{\code{ukb_gen_write_bgenie}}
}
