% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_util.R
\name{ukb_md5}
\alias{ukb_md5}
\alias{ukb_unpack}
\alias{ukb_conv}
\title{UKB MD5 Checksum}
\usage{
ukb_md5(file, ...)

ukb_unpack(file, key, ...)

ukb_conv(
  file,
  type = c("r", "docs", "csv", "sas", "stata", "lims", "bulk", "txt"),
  ...
)
}
\arguments{
\item{file}{name of file to run utility on}

\item{...}{additional arguments to pass to
\code{\link{ukb_util_path}}}

\item{key}{file to key to unpack/decrypt file}

\item{type}{type of conversion to do}
}
\value{
A character string
}
\description{
UKB MD5 Checksum
}
